+++ src/lib/Components/Bidding/Screens/RegistrationResult.tsx
@@ -38,9 +38,9 @@ const registrationCompleteMessage = {
 const registrationPendingMessage = {
   title: "Registration pending",
   description:
-    "Your registration status is pending.\n" +
-    "Please contact [support@artsy.net](mailto:support@artsy.net) for\n" +
-    "more information.",
+    "You will receive an email from an Artsy specialist once your registration has been confirmed.\n" +
+    "\n" +
+    "In the meantime, you can still view works and watch lots that you’re interested in.",
 }
 
 const registrationErrorMessage = {
@@ -104,7 +104,7 @@ export class RegistrationResult extends React.Component<RegistrationResultProps,
         <Container mt={6}>
           <View>
             <Flex alignItems="center">
-              <Icon20 source={Icons[status]} />
+              {status !== RegistrationStatus.RegistrationStatusPending && <Icon20 source={Icons[status]} />}
               <Title mt={2} mb={4}>
                 {title}
               </Title>